# docs/swagger.yaml
basePath: /
definitions:
  schemas.AllCollections:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  schemas.AllCollectionsResp:
    properties:
      collections:
        items:
          $ref: '#/definitions/schemas.AllCollections'
        type: array
    type: object
  schemas.CardsByCollectionID:
    properties:
      answer:
        type: string
      id:
        type: integer
      question:
        type: string
    type: object
...
  schemas.CreateCollectionReq:
    properties:
      description:
        type: string
      name:
        type: string
    required:
    - name
    type: object
  schemas.CreateCollectionResp:
    properties:
      createdAt:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
...
  schemas.GetCollectionByIDResp:
    properties:
      createdAt:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
...
  schemas.PatchCollectionReq:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    required:
    - id
    type: object
...
  schemas.TrainSchemaResp:
    properties:
      cards:
        items:
          $ref: '#/definitions/schemas.CardsByCollectionID'
        type: array
    type: object
...
  schemas.UpdateCollectionReq:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    required:
    - id
    - name
    type: object
  schemas.UpdateCollectionResp:
    properties:
      createdAt:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
...
  util.ErrorResponse:
    properties:
      error:
        type: string
    type: object
info:
  contact: {}
  description: REST API for flashcard learning platform.
  title: Card Backend API
  version: "1.1"
paths:
...
  /collections/:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.AllCollectionsResp'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all collections
      tags:
      - collections
    post:
      consumes:
      - application/json
      parameters:
      - description: collection payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schemas.CreateCollectionReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/schemas.CreateCollectionResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create collection
      tags:
      - collections
  /collections/{collectionID}/:
    delete:
      parameters:
      - description: collection id
        in: path
        name: collectionID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete collection
      tags:
      - collections
    get:
      parameters:
      - description: collection id
        in: path
        name: collectionID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.GetCollectionByIDResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get collection by ID
      tags:
      - collections
    patch:
      consumes:
      - application/json
      parameters:
      - description: collection id
        in: path
        name: collectionID
        required: true
        type: integer
      - description: collection payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schemas.PatchCollectionReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.UpdateCollectionResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/util.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Partial update collection
      tags:
      - collections
    put:
      consumes:
      - application/json
      parameters:
      - description: collection id
        in: path
        name: collectionID
        required: true
        type: integer
      - description: collection payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schemas.UpdateCollectionReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.UpdateCollectionResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/util.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Full update collection
      tags:
      - collections
...
  /collections/{collectionID}/train:
    get:
      parameters:
      - description: collection id
        in: path
        name: collectionID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.TrainSchemaResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/util.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Train cards
      tags:
      - collections
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"